# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PLoT Engine API
  version: 1.0.0
  description: |
    PLoT Engine v1 - Deterministic, auditable probabilistic models with trust signals.
    
    **Trust Signals**: All /v1 responses include Model Card, Confidence badge, and Explain-Δ.
    
    **Demo Mode**: Add `?demo=1` or header `X-Demo: 1` for deterministic sample payloads.
    When present, demo mode short-circuits request validation and returns a canonical demo payload.
    
    **Feature Flags**:
    - FEATURE_STREAM=1 enables real SSE /stream route (test-only otherwise)
    - AUTH_ENABLED=1 requires Authorization: Bearer <token> on certain routes
    
    **Legacy Routes**: /draft-flows, /critique, /improve (v0) remain for backwards compatibility.
  contact:
    name: PLoT Engine
    url: https://github.com/Talchain/plot-lite-service
    email: devnull@talchain.ai
servers:
  - url: http://127.0.0.1:4311
    description: Local development
  - url: https://api.olumi.ai
    description: Production (future)
...
paths:
...
  /v1/intervene:
    post:
      summary: Causal interventions (do-operator)
      description: |
        Perform causal interventions and estimate counterfactual effects.
        Returns baseline, counterfactual, and delta with identifiability check.
      operationId: intervene
      tags: [v1, inference]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [graph, do]
              properties:
                graph:
                  type: object
                  required: [nodes, edges]
                  properties:
                    nodes:
                      type: array
                      items:
                        type: object
                        required: [id]
                        properties:
                          id:
                            type: string
                          label:
                            type: string
                    edges:
                      type: array
                seed:
                  type: integer
                  example: 4242
                actions:
                  type: array
                  description: Causal interventions via do-operator
                  items:
                    type: object
                    required: [node_id, value]
                    properties:
                      node_id:
                        type: string
                      value:
                        type: number
                flags:
                  type: object
                  properties:
                    scm_lite:
                      type: integer
                      enum: [0, 1]
            example:
              graph:
                nodes:
                  - {id: "X", label: "Cause"}
                  - {id: "Y", label: "Effect"}
                edges:
                  - {from: "X", to: "Y", weight: 0.5}
              seed: 4242
              actions:
                - {node_id: "X", value: 0.8}
              flags: {scm_lite: 1}
      responses:
        '200':
          description: Intervention results
          content:
            application/json:
              schema:
                type: object
                properties:
                  schema:
                    type: string
                    example: "intervene.v1"
                  baseline:
                    type: object
                    properties:
                      summary:
                        type: object
                        properties:
                          p10: {type: number}
                          p50: {type: number}
                          p90: {type: number}
                  counterfactual:
                    type: object
                    properties:
                      summary:
                        type: object
                        properties:
                          p10: {type: number}
                          p50: {type: number}
                          p90: {type: number}
                  delta:
                    type: object
                    properties:
                      p10: {type: number}
                      p50: {type: number}
                      p90: {type: number}
                  explain:
                    type: object
                    properties:
                      provenance:
                        type: string
                        example: "do-operator"
                      top_drivers:
                        type: array
                        items:
                          type: object
                          properties:
                            node_id: {type: string}
                            contribution: {type: number}
                            sign: {type: string, enum: ["+", "-"]}
                  model_card:
                    type: object
                    properties:
                      actions_hash: {type: string}
                      seed: {type: integer}
                      flags_on: {type: array, items: {type: string}}
                  response_hash:
                    type: string
              example:
                schema: "intervene.v1"
                baseline:
                  summary: {p10: 0.819, p50: 0.924, p90: 1.109}
                counterfactual:
                  summary: {p10: 0.915, p50: 1.044, p90: 1.253}
                delta: {p10: 0.096, p50: 0.120, p90: 0.144}
                explain:
                  provenance: "do-operator"
                  top_drivers:
                    - {node_id: "X", contribution: 80, sign: "+"}
                model_card:
                  actions_hash: "a1b2c3d4e5f6g7h8"
                  seed: 4242
                  flags_on: ["scm_lite"]
                response_hash: "1234567890abcdef"
        '400':
          description: Bad input or not identifiable
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      type:
                        type: string
                      message:
                        type: string
              examples:
                bad_input:
                  value:
                    schema: "error.v1"
                    code: "BAD_INPUT"
                    message: "Invalid node_ids in actions: Z"
                missing_actions:
                  value:
                    schema: "error.v1"
                    code: "BAD_INPUT"
                    message: "actions array required with at least one intervention"
  /v1/evidence:
    post:
      summary: Evidence & Priors
      description: |
        Apply reference-class priors to influence beliefs.
        Returns adjusted inference with baseline comparison.
        Note: v1 uses simple weighted average; full Bayesian update in future versions.
      operationId: evidence
      tags: [v1, inference]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [graph, priors]
              properties:
                graph:
                  type: object
                  required: [nodes, edges]
                  properties:
                    nodes:
                      type: array
                      items:
                        type: object
                        required: [id]
                        properties:
                          id:
                            type: string
                          label:
                            type: string
                    edges:
                      type: array
                seed:
                  type: integer
                  example: 4242
                priors:
                  type: array
                  items:
                    type: object
                    required: [node_id, mean, variance]
                    properties:
                      node_id:
                        type: string
                      mean:
                        type: number
                      variance:
                        type: number
                      source:
                        type: string
                        example: "refclass:xyz"
            example:
              graph:
                nodes:
                  - {id: "A", label: "Node A"}
                edges: []
              seed: 4242
              priors:
                - {node_id: "A", mean: 0.8, variance: 0.05, source: "refclass:xyz"}
      responses:
        '200':
          description: Evidence results
          content:
            application/json:
              schema:
                type: object
                properties:
                  schema:
                    type: string
                    example: "evidence.v1"
                  result:
                    type: object
                    properties:
                      summary:
                        type: object
                        properties:
                          p10:
                            type: number
                          p50:
                            type: number
                          p90:
                            type: number
                      baseline:
                        type: object
                        properties:
                          p50:
                            type: number
                      adjustment:
                        type: object
                        properties:
                          delta:
                            type: number
                          direction:
                            type: string
                            enum: [increase, decrease, none]
                  top_drivers:
                    type: array
                  meta:
                    type: object
                    properties:
                      seed:
                        type: integer
                      inference_mode:
                        type: string
                      prior_alpha:
                        type: number
                      limitations:
                        type: string
              example:
                schema: "evidence.v1"
                result:
                  summary: {p10: 0.524, p50: 0.724, p90: 0.924}
                  baseline: {p50: 0.624}
                  adjustment: {delta: 0.1, direction: "increase"}
                top_drivers:
                  - {node_id: "A", contribution: 80, sign: "+", source: "refclass:xyz"}
                meta: 
                  seed: 4242
                  inference_mode: "model_based"
                  prior_alpha: 0.3
                  limitations: "v1: Simple weighted average. Full Bayesian update in future versions."
        '400':
          description: Bad input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      type:
                        type: string
                      message:
                        type: string
              examples:
                invalid_prior:
                  value:
                    schema: "error.v1"
                    code: "BAD_INPUT"
                    message: "Invalid node_ids in priors: Z"
...